+++ extensions/dbm_dashboard/dbm_dashboard_EXT/data/globalVars.json
@@ -1,0 +1,1 @@
+{}
