+++ extensions/dbm_dashboard/dbm_dashboard_EXT/views/customRoute.ejs
