+++ extensions/dbm_dashboard/dbm_dashboard_EXT/views/homePage.ejs
@@ -1,0 +1,90 @@
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title><%= client.user.username %></title>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
+    <link href="<%= theme %>" rel="stylesheet">
+    <link rel="shortcut icon" href="<%= client.user.avatarURL || 'https://discordapp.com/assets/322c936a8c8be1b803cd94861bdfa868.png' %>"/>
+    <meta name="theme-color" content="#7289DA" />
+    <meta name="description" content="<%= config.introText %>">
+</head>
+
+<body>
+    <div class="d-flex flex-column flex-md-row align-items-center p-3 px-md-4 mb-3 navHeader">
+        <h5 class="my-0 mr-md-auto headerBrand"><%= client.user.username %></h5>
+        <nav class="my-2 my-md-0 mr-md-3 headerBrandNavButtons">
+            <% for (i = 0; i < config.navItems.length; i++) { %>
+            <a class="p-2" href="<%= config.navItems[i].link %>"><%= config.navItems[i].name %></a>
+            <% } %>
+        </nav>
+    </div>
+    <section class="jumbotron text-center lightBackground">
+        <div class="container">
+            <img style="border-radius: 50%; width: 100px; height: 100px;"
+                src="<%= client.user.avatarURL || 'https://discordapp.com/assets/322c936a8c8be1b803cd94861bdfa868.png' %>"
+                alt="">
+            <h1 class="jumbotron-heading" style="color: #7289DA;"><%= client.user.username %></h1>
+            <p class="lead "><%= config.introText %></p>
+            <p>
+                <a href="<%= config.inviteLink %>" class="btn btn-primary my-2"
+                    style="background-color: #7289DA; width: 132px;">Invite Bot</a>
+                <a href="<%= config.supportServer %>" class="btn btn-primary my-2"
+                    style="background-color: #7289DA; width: 132px;">Support Server</a>
+            </p>
+        </div>
+    </section>
+    <div style="height: 150px; overflow: hidden; margin-top: -100px;">
+        <svg viewBox="0 0 500 150" preserveAspectRatio="none" style="height: 100%; width: 100%;">
+            <path d="M0.00,49.98 C149.99,150.00 349.20,-49.98 500.00,49.98 L500.00,150.00 L0.00,150.00 Z"
+                style="stroke: none;fill: #23272A;"></path>
+        </svg>
+    </div>
+    <section class="darkBackground">
+        <div class="container text-center darkBackground features">
+            <div class="row text-center">
+                <div class="col-md-6">
+                    <p class="featuresHeader"><%= config.featureOne.name %></p>
+                    <p class="features"><%= config.featureOne.description %></p>
+                </div>
+                <div class="col-md-6 d-none d-lg-block"></div>
+                <div class="col-md-6 d-none d-lg-block"></div>
+                <div class="col-md-6">
+                    <p class="featuresHeader"><%= config.featureTwo.name %></p>
+                    <p><%= config.featureTwo.description %></p>
+                </div>
+                <div class="col-md-6">
+                    <p class="featuresHeader"><%= config.featureThree.name %></p>
+                    <p class="features"><%= config.featureThree.description%></p>
+                </div>
+                <div class="col-md-6 d-none d-lg-block"></div>
+                <div class="col-md-6 d-none d-lg-block"></div>
+                <div class="col-md-6">
+                    <p class="featuresHeader"><%= config.featureFour.name %></p>
+                    <p><%= config.featureFour.description %></p>
+                </div>
+            </div>
+        </div>
+        <div style="height: 150px; overflow: hidden;">
+            <svg viewBox="0 0 500 150" preserveAspectRatio="none" style="height: 100%; width: 100%;">
+                <path d="M0.00,49.98 C149.99,150.00 349.20,-49.98 500.00,49.98 L500.00,150.00 L0.00,150.00 Z"
+                    style="stroke: none;fill: #2C2F33;"></path>
+            </svg>
+        </div>
+    </section>
+    <section id="lightBackground" style="margin-top: -60px;">
+        <footer class="page-footer font-small">
+            <div class="container text-center text-md-left mt-5">
+                <div class="row mt-3">
+                    <h6 class="text-uppercase font-weight-bold featuresHeaderFooter"><%= client.user.username %></h6>
+                    <hr class="deep-purple accent-2 mb-4 mt-0 d-inline-block mx-auto" style="width: 100%;">
+                    <p style="color: white; max-width: 900px;"><%= config.footerText %></p>
+                </div>
+            </div>
+        </footer>
+    </section>
+</body>
+
+</html>
