--- extensions/dbm_dashboard/dbm_dashboard_extension/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "dbm-dashboard",
-  "version": "1.0.0",
-  "author": "Great Plains Modding",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "chalk": "^3.0.0",
-    "figlet": "^1.2.4",
-    "cookie-parser": "^1.4.4",
-    "ejs": "^3.0.1",
-    "express": "^4.17.1",
-    "express-session": "^1.17.0",
-    "fs": "0.0.1-security",
-    "passport": "^0.4.1",
-    "passport-discord": "^0.1.3",
-    "path": "^0.12.7"
-  }
-}
